+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   - Example: 10.2.1.4 is the 5th version that supports khiops 10.2.1.
 - Internals: Changes in *Internals* sections are unlikely to be of interest for data scientists.
 
+## Unreleased
+
+## Fixed
+- (`core`) Samples dir path construction when HOME is a remote path
+
 ## 11.0.0.1 - 2026-01-14
 
 ### Fixed