+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   - Example: 10.2.1.4 is the 5th version that supports khiops 10.2.1.
 - Internals: Changes in *Internals* sections are unlikely to be of interest for data scientists.
 
-## Unreleased
+## 11.0.0.3 - 2026-03-06
 
 ### Added
 - (`sklearn`) `n_feature_parts` parameter to the supervised estimators